[dagster-shared] move check & record
