+++ python_modules/dagster/dagster_tests/general_tests/test_serdes.py
@@ -9,7 +9,6 @@ from typing import AbstractSet, Any, NamedTuple, Optional, Union  # noqa: UP035
 import dagster._check as check
 import pydantic
 import pytest
-from dagster._check.functions import CheckError
 from dagster._model import DagsterModel
 from dagster._record import IHaveNew, record, record_custom
 from dagster._serdes.errors import DeserializationError, SerdesUsageError, SerializationError
@@ -31,6 +30,7 @@ from dagster._serdes.serdes import (
 )
 from dagster._serdes.utils import hash_str
 from dagster._utils.cached_method import cached_method
+from dagster_shared.check import CheckError
 
 
 def test_deserialize_value_ok():
