+++ python_modules/libraries/dagster-components/dagster_components_tests/scaffoldable_tests/test_decorator.py
@@ -1,6 +1,6 @@
 import pytest
-from dagster._check.functions import CheckError
 from dagster_components.scaffold import Scaffolder, get_scaffolder, has_scaffolder, scaffold_with
+from dagster_shared.check import CheckError
 
 
 # Example usage:
