+++ python_modules/libraries/dagster-shared/MANIFEST.in
@@ -1,3 +1,4 @@
 include README.md
 include LICENSE
 include dagster_shared/py.typed
+recursive-include dagster_shared *.md
