+++ python_modules/libraries/dagster-shared/dagster_shared_tests/conftest.py
@@ -1,0 +1,8 @@
+import os
+
+import pytest
+
+
+@pytest.fixture(scope="session", autouse=True)
+def enable_defensive_checks():
+    os.environ["DAGSTER_RECORD_DEFENSIVE_CHECKS"] = "true"
